pyproject.toml
[tool.poetry]
name = "hvat"
version = "0-dev"
description = ""
authors = ["Eirik W. Talberg <eirik@taldev.no>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
hvac = "^2.0.0"
click = "^8.1.7"
pyyaml = "^6.0.1"


[tool.poetry.group.dev.dependencies]
flake8-pyproject = "^1.2.3"
pytest = "^7.4.3"
flake8 = "^6.1.0"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.flake8]
max-line-length = 120

[tool.pytest.ini_options]
junit_suite_name = "demoapp"
addopts = "--junitxml=reports/pytest.xml --cov=. --cov-report term --cov-report xml:reports/coverage.xml"

[tool.coverage.run]
omit = [ 
    "tests/**",
    "hvat/main.py"
]